feat: use int64_t instead int

+++ contracts/graphenelib/global.h
@@ -22,11 +22,11 @@ extern "C" {
     // get current transaction size of current contract calling
     int transaction_size();
     // return current transaction expiration
-    int expiration();
+    uint64_t expiration();
     // return ref block num(block_id.hash[0])
     // eg. "block_id": "00000fa00f4dd71912f56dd6e23f03bf2af87be5" --> 00000fa0
     int tapos_block_num();
     // return ref block prefix(block_id.hash[1])
     //eg. "block_id": "00000fa00f4dd71912f56dd6e23f03bf2af87be5" --> 0f4dd719
-    int tapos_block_prefix();
+    uint64_t tapos_block_prefix();
 }
